clock-make transparent calendars possible
Add support for transparent calendar windows, the window background color can be set to any color and alpha value or made transparent with the #MatePanelPopupWindow named widget selector in the gtk theme. Defaults to theme background color if not set, same as before. This is used only for the calendar window right now