panel-context-menu.c: custom menu theme support
Make the panel context menu follow the .mate-panel-menu-bar .menu selector in the gtk theme, and the  panel context menu window follow the .mate-panel-menu-bar selector. Use the former to set a custom menu theme, the latter to alter or remove shadows from applet menus. Include robust support for transparent themes even with gtk does not detect a compositing window manager so long as compositing is enabled.